backup & restore stable plugin files
When a beta branch is activated, the original files are backed up to
`wp-content/podlove-beta-backup/<plugin-slug>`. When it’s deactivated,
the original files are copied back to restore the “stable” file state.

This *may* cause issues when migrations introduce backwards
incompatible changes (like schema migrations). This *should* then be
stated in beta branch notes.